test: test fixtures refactoring
